+++ scenarios/astronomical-timer/astronomical-timer.mod.js
@@ -13,6 +13,7 @@ var Logger = require('logger.mod').Logger;
 var SunCalc = require('suncalc.mod');
 var aTable = require('table-handling-actions.mod');
 var constants = require('constants.mod');
+var durationReverse = require('duration-reverse.mod');
 var isControlTypeValid =
   require('scenarios-general-helpers.mod').isControlTypeValid;
 
@@ -22,7 +23,6 @@ var log = new Logger(loggerFileLabel);
 var DAY_NAMES = constants.DAY_NAMES;
 var DAY_NAME_TO_NUMBER = constants.DAY_NAME_TO_NUMBER;
 var VALID_DAYS = constants.VALID_DAYS;
-var FULL_DAYS = constants.FULL_DAYS;
 
 var ASTRO_EVENT_NAMES = {
   sunrise: { en: 'Sunrise', ru: 'Восход' },
@@ -40,21 +40,9 @@ var ASTRO_EVENT_NAMES = {
 };
 
 var MS_PER_MINUTE = constants.MS_PER_MINUTE;
-var MS_PER_SECOND = constants.MS_PER_SECOND;
-var MS_PER_HOUR = constants.MS_PER_HOUR;
 var MAX_DAYS_AHEAD = 365;
 var OFFSET_MIN_MIN = -720; // -12 hours
 var OFFSET_MAX_MIN = 720; // +12 hours
-var MAX_DURATION_MS = 12 * MS_PER_HOUR;
-
-// Actions reversed after the delay: toggle flips back, setValue/setText/
-// setColor apply reverseValue
-var REVERSIBLE_ACTIONS = {
-  toggle: true,
-  setValue: true,
-  setText: true,
-  setColor: true,
-};
 
 /**
  * @typedef {Object} Coordinates
@@ -313,30 +301,10 @@ AstronomicalTimerScenario.prototype.validateCfg = function (cfg) {
   }
 
   // Validate the optional turn-off delay
-  if (cfg.duration && cfg.duration.value !== 0) {
-    var validUnits = ['hours', 'minutes', 'seconds'];
-    if (validUnits.indexOf(cfg.duration.unit) === -1) {
-      log.error(
-        'Astronomical Timer validation error: duration.unit must be hours, minutes or seconds'
-      );
-      return false;
-    }
-    if (
-      typeof cfg.duration.value !== 'number' ||
-      cfg.duration.value < 0 ||
-      cfg.duration.value % 1 !== 0
-    ) {
-      log.error(
-        'Astronomical Timer validation error: duration.value must be a non-negative integer'
-      );
-      return false;
-    }
-    if (durationToMs(cfg.duration) > MAX_DURATION_MS) {
-      log.error(
-        'Astronomical Timer validation error: turn-off delay must not exceed 12 hours'
-      );
-      return false;
-    }
+  var durationError = durationReverse.validateDuration(cfg);
+  if (durationError) {
+    log.error('Astronomical Timer validation error: {}', durationError);
+    return false;
   }
 
   // Check that the event actually exists in the next MAX_DAYS_AHEAD days
@@ -392,7 +360,7 @@ function addCustomControlsToVirtualDevice(self, cfg) {
 
   // Add next execution time display control
   var nextExecution = calculateAndCacheEventTime(self, cfg);
-  var nextExecutionText = formatNextExecution(nextExecution);
+  var nextExecutionText = durationReverse.formatNextExecution(nextExecution);
   self.vd.devObj.addControl('next_execution', {
     title: {
       en: 'Next execution',
@@ -457,7 +425,7 @@ function addCustomControlsToVirtualDevice(self, cfg) {
   }
 
   // Add turn-off time display only when the turn-off timer is in use
-  if (usesTurnOffTimer(cfg)) {
+  if (durationReverse.usesTurnOffTimer(cfg)) {
     self.vd.devObj.addControl('return_time', {
       title: {
         en: 'Turns off at',
@@ -598,7 +566,7 @@ function updateEventTimeAndDisplay(self, cfg) {
   // to avoid unnecessary writes to MQTT on every cron tick
   if (self.ctx.cachedNextExecutionMs !== prevNextExecutionMs) {
     dev[self.genNames.vDevice + '/next_execution'] =
-      formatNextExecution(nextExecution);
+      durationReverse.formatNextExecution(nextExecution);
 
     if (cfg.eventSettings.offset !== 0) {
       if (nextExecution) {
@@ -638,38 +606,6 @@ function formatCurrentTime() {
   var currentDay = DAY_NAMES[currentDayNum] || 'Unknown';
 
   return currentDay + ' ' + currentDate + ' ' + currentTime;
-}
-
-/**
- * Format date for display
- * @param {Date} date
- * @returns {string} - Formated dateTime
- */
-function formatNextExecution(date) {
-  if (!date) {
-    return '--:--';
-  }
-
-  var dayName = FULL_DAYS[date.getDay()];
-  var day = ('0' + date.getDate()).slice(-2);
-  var month = ('0' + (date.getMonth() + 1)).slice(-2);
-  var year = date.getFullYear();
-  var hours = ('0' + date.getHours()).slice(-2);
-  var minutes = ('0' + date.getMinutes()).slice(-2);
-
-  return (
-    dayName +
-    ' ' +
-    year +
-    '-' +
-    month +
-    '-' +
-    day +
-    ' ' +
-    hours +
-    ':' +
-    minutes
-  );
 }
 
 /**
@@ -739,102 +675,6 @@ function getNextExecutionTime(cfg) {
   return null;
 }
 
-// Convert a duration object {unit, value} to milliseconds
-function durationToMs(duration) {
-  if (duration.unit === 'hours') {
-    return duration.value * MS_PER_HOUR;
-  }
-  if (duration.unit === 'seconds') {
-    return duration.value * MS_PER_SECOND;
-  }
-  return duration.value * MS_PER_MINUTE;
-}
-
-function isDurationEnabled(cfg) {
-  return !!(cfg.duration && cfg.duration.value >= 1);
-}
-
-function hasReversibleControls(cfg) {
-  for (var i = 0; i < cfg.outControls.length; i++) {
-    if (REVERSIBLE_ACTIONS[cfg.outControls[i].behaviorType]) {
-      return true;
-    }
-  }
-  return false;
-}
-
-// A turn-off timer is used only when a delay is set and something is reversible
-function usesTurnOffTimer(cfg) {
-  return isDurationEnabled(cfg) && hasReversibleControls(cfg);
-}
-
-// Reverse reversible controls - toggle flips, setValue/setText/setColor apply
-// reverseValue. Empty reverseValue is skipped
-function executeReverse(cfg) {
-  for (var i = 0; i < cfg.outControls.length; i++) {
-    var outControl = cfg.outControls[i];
-    var behaviorType = outControl.behaviorType;
-    if (!REVERSIBLE_ACTIONS[behaviorType]) {
-      continue;
-    }
-
-    var curCtrlName = outControl.control;
-    try {
-      var actualValue = dev[curCtrlName];
-      var newCtrlValue;
-
-      if (behaviorType === 'toggle') {
-        newCtrlValue = aTable.actionsTable.toggle.handler(actualValue);
-      } else {
-        var reverseValue = outControl.reverseValue;
-        if (reverseValue === undefined || reverseValue === '') {
-          continue;
-        }
-        newCtrlValue = aTable.actionsTable[behaviorType].handler(
-          actualValue,
-          reverseValue
-        );
-      }
-
-      dev[curCtrlName] = newCtrlValue;
-    } catch (error) {
-      log.error(
-        'Failed to reverse control {}: {}',
-        curCtrlName,
-        error.message || error
-      );
-    }
-  }
-}
-
-function setReturnTimeDisplay(self, text) {
-  if (self.vd.devObj.getControl('return_time')) {
-    dev[self.genNames.vDevice + '/return_time'] = text;
-  }
-}
-
-function cancelOffTimer(self) {
-  if (self.ctx.offTimerId !== null) {
-    clearTimeout(self.ctx.offTimerId);
-    self.ctx.offTimerId = null;
-  }
-  setReturnTimeDisplay(self, '--:--');
-}
-
-// Arm the timer that reverses controls after the delay
-function armOffTimer(self, cfg) {
-  var delayMs = durationToMs(cfg.duration);
-  var returnDate = new Date(Date.now() + delayMs);
-
-  self.ctx.offTimerId = setTimeout(function turnOffHandler() {
-    self.ctx.offTimerId = null;
-    executeReverse(cfg);
-    setReturnTimeDisplay(self, '--:--');
-  }, delayMs);
-
-  setReturnTimeDisplay(self, formatNextExecution(returnDate));
-}
-
 /**
  * Handler for astronomical timer trigger
  * @param {AstronomicalTimerScenario} self - Reference to the AstronomicalTimerScenario instance
@@ -851,8 +691,8 @@ function astroHandler(self, cfg) {
   }
 
   // Cancel any pending timer so the previous window cannot reverse mid-way
-  if (usesTurnOffTimer(cfg)) {
-    cancelOffTimer(self);
+  if (durationReverse.usesTurnOffTimer(cfg)) {
+    durationReverse.cancelOffTimer(self, self.ctx);
   }
 
   // Execute all configured actions
@@ -887,8 +727,8 @@ function astroHandler(self, cfg) {
   }
 
   // Arm the turn-off timer when a delay is set and there is something to reverse
-  if (usesTurnOffTimer(cfg)) {
-    armOffTimer(self, cfg);
+  if (durationReverse.usesTurnOffTimer(cfg)) {
+    durationReverse.armOffTimer(self, cfg, self.ctx);
   }
 
   log.debug('Astro timer actions completed for: {}', self.idPrefix);
@@ -1002,18 +842,25 @@ function createTimeUpdateRule(self) {
   return true;
 }
 
+/**
+ * Creates cleanup rule that reverses controls and cancels the turn-off timer
+ * when the scenario is disabled. Not registered via self.addRule so it keeps
+ * working after the scenario stops.
+ * @param {AstronomicalTimerScenario} self - Reference to the AstronomicalTimerScenario instance
+ * @param {AstronomicalTimerConfig} cfg - Configuration object
+ * @returns {boolean} True if rule created successfully, false otherwise
+ */
 function createDisableRule(self, cfg) {
   var disableRuleId = defineRule(self.genNames.ruleDisable, {
     whenChanged: [self.genNames.vDevice + '/rule_enabled'],
     then: function disableCleanupHandler(newValue) {
       if (!newValue && self.ctx.offTimerId !== null) {
-        executeReverse(cfg);
-        cancelOffTimer(self);
+        durationReverse.executeReverse(cfg);
+        durationReverse.cancelOffTimer(self, self.ctx);
       }
     },
   });
 
-  // This rule not disable when user use switch in virtual device
   if (!disableRuleId) {
     log.error('Failed to create disable rule');
     return false;
@@ -1044,7 +891,10 @@ function createRules(self, cfg) {
   }
 
   // Only needed when a turn-off timer can be active
-  if (usesTurnOffTimer(cfg) && !createDisableRule(self, cfg)) {
+  if (
+    durationReverse.usesTurnOffTimer(cfg) &&
+    !createDisableRule(self, cfg)
+  ) {
     return false;
   }
